Latest checkin.

+++ src/heuristics.c
@@ -2,12 +2,12 @@
 #include <math.h>
 #include "heuristics.h"
 
-unsigned int manhattan_distance( const coordinate_t *c1, const coordinate_t *c2 )
+unsigned long manhattan_distance( const coordinate_t *c1, const coordinate_t *c2 )
 {
-	return (unsigned int) abs( (int) (c1->x - c2->x) ) + abs( (int) (c1->y - c2->y) );
+	return (unsigned long) abs( (long) (c1->x - c2->x) ) + abs( (long) (c1->y - c2->y) );
 }
 
-unsigned int euclidean_distance( const coordinate_t *c1, const coordinate_t *c2 )
+unsigned long euclidean_distance( const coordinate_t *c1, const coordinate_t *c2 )
 {
-	return (unsigned int) sqrt( (float) ((c1->x - c2->x) * (c1->x - c2->x) + (c1->y - c2->y) * (c1->y - c2->y)) );
+	return (unsigned long) sqrt( (float) ((c1->x - c2->x) * (c1->x - c2->x) + (c1->y - c2->y) * (c1->y - c2->y)) );
 }
